PR-05: extract control/{mod,client}.rs
First nested-directory module. Behaviour-preserving relocation:
- control/mod.rs: ControlCommand + impl, ControlDispatch.
- control/client.rs: send_control_command*, BroadcastEntryReport,
  BroadcastReport, enumerate_daemon_names,
  send_control_command_broadcast, plus BROADCAST_PER_CALL_TIMEOUT.

main.rs: 5945 -> 5791 LOC. args.rs updated to use
crate::control::ControlCommand (was super::ControlCommand from PR-04a).

Verification: cargo build clean; cargo test --bin kanata-switcher
261 passed / 0 failed.
